Increase max-nested-callbacks to 3
Increased due to good code hitting the limit in the following use cases:
- Interoperating with third-party code (e.g., tests, task runners)
- Promise chains

Code that operates at 2 different abstraction levels, which is the max
allowed by max-depth, sometimes is passed as a callback to stuff like
test frameworks or promise chains. In those cases, the rule
max-nested-callbacks detects a callback at depth 3. However, the
conceptual "depth" of the code is still 2. Due to this, usually there's
no way of refactoring the code without hurting legibility.