Remove implicit dependencies on truthy defaults
That way the config still works consistently when a newer versions of
eslint changes a default.

Options that default to false are omited because it's redundant to say
that we don't want to use a given option. If eslint were to introduce a
new option that defaults to true, that's a design error on them and the
above paragraph doesn't apply.