Use Prettier's defaults
This way one can configure Prettier without overriding rules from this
config, as it's designed not to conflict with Prettier.

The only exception is the printWidth option, which must match the
value set by the max-len rule. Deviating from the standard width of
80 characters is advised against, more so when using a formatter like
Prettier, hence it should rarely be a problem.