Remove for loops from no-restricted-syntax
There are some legitimate use cases for the `for in` loop, such as:
- Lookups of keys in the prototype chain
- Early exits when iterating over keys, for perfomance reasons, instead
of `Object.keys`, which has to lookup all keys and create an array

Also, other rules like guard-for-in already make for in loops safer.

We find no reason to use the plain for loop and the for of loop,
as you can use higher order functions or while loops instead. However,
that is the most controversial rule of this config, hence the most often
disabled one. This is the major cause of friction when using the config.

Since no-restricted-syntax forbids other constructs, disabling it
because you want to use some form of for loop has the unfortunate side
effect of also disabling linting of everything else in the rule, which
is bad since it enables greater evils than for loops.